+++ projects/ng-rule-editor/src/lib/rule-editor.service.spec.ts
@@ -53,12 +53,19 @@ describe('RuleEditorService', () => {
     service.import('http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-calculatedExpression',
       phq9, '/39156-5');
     const original = phq9;
+
+    // Check to make sure we start with the expected number of extensions on the target item
+    const INITIAL_EXTENSION_COUNT = 4;
+    const CONTEXT_INDEX = 9;
+    expect(original.item[CONTEXT_INDEX].extension.length).toEqual(INITIAL_EXTENSION_COUNT);
+
     const withScores = service.exportSumOfScores();
 
-    // TODO assert
+    expect(withScores['item'][CONTEXT_INDEX].extension.length).toEqual(INITIAL_EXTENSION_COUNT + 11);
 
     const removedScores = service.removeSumOfScores(withScores);
 
+    expect(removedScores['item'][CONTEXT_INDEX].extension.length).toEqual(INITIAL_EXTENSION_COUNT);
     expect(removedScores).toEqual(phq9);
   });
 });